Renamed the outlook-sendmail command to mail send
