+++ src/o365/outlook/commands.ts
@@ -1,5 +1,6 @@
 const prefix: string = 'outlook';
 
 export default {
-  OUTLOOK_SENDMAIL: `${prefix} sendmail`
+  OUTLOOK_SENDMAIL: `${prefix} sendmail`,
+  OUTLOOK_MAIL_SEND: `${prefix} mail send`
 };
